Extracted filter fields from expositions header to separate components
